создана функция createPreviews
